Extract Parsing and Serialization into their own files
